modify calculator
